Move many tooltips
