--- site/src/components/Tooltips/index.ts
@@ -1,4 +1,0 @@
-export { AuditHelpTooltip } from "./AuditHelpTooltip"
-export { WorkspaceOutdatedTooltip } from "./WorkspaceOutdatedTooltip"
-export { UserRoleHelpTooltip } from "./UserRoleHelpTooltip"
-export { WorkspaceHelpTooltip } from "./WorkspaceHelpTooltip"
