+++ site/src/components/UsersTable/UserRoleHelpTooltip.tsx
@@ -5,7 +5,7 @@ import {
   HelpTooltipLinksGroup,
   HelpTooltipText,
   HelpTooltipTitle,
-} from "./HelpTooltip"
+} from "components/HelpTooltip/HelpTooltip"
 import { docs } from "utils/docs"
 
 export const Language = {
